Introduce new Rank API.
Supports all ranks, package ranks and player ranks. Supports sorting of
the ranks as well as methods to return the highest rank of an array.
Includes a basic method to return the symbol from a the API’s response
string for a rank.

Introduces a rank getter to the Player class.